requirements.txt
tensorflow==2.15.0
Flask==3.0.3
pandas==2.1.4
opencv-python== 4.9.0.80
scikit-learn==1.2.2
pillow==9.5.0
numpy==1.26.4
seaborn==0.12.2
matplotlib==3.7.5
gunicorn==20.1.0

